+++ src/Servant/Pagination.hs
@@ -1,75 +1,278 @@
-{-# LANGUAGE TypeFamilies #-}
+-- | Opinionated Pagination Helpers for Servant APIs
+--
+--
+-- Client can provide a `Range` header with their request with the following format
+--
+-- > Range: <field> [<value>][; offset <o>][; limit <l>][; order <asc|desc>]
+--
+-- Available ranges are declared using type-level list of accepted fields, bound to a given
+-- resource and type using the 'HasPagination' type-class. The library provides unobtrusive
+-- types and abstract away all the plumbing to hook that on an existing Servant API.
+--
+-- The 'IsRangeType' constraints summarize all constraints that must apply to a possible field
+-- and heavily rely on the 'Web.Internal.FromHttpApiData' and 'Web.Internal.ToHttpApiData'.
+--
+-- > $ curl -v http://localhost:1337/colors -H 'Range: name; limit 10'
+-- >
+-- > > GET /colors HTTP/1.1
+-- > > Host: localhost:1337
+-- > > User-Agent: curl/7.47.0
+-- > > Accept: */*
+-- > >
+-- > < HTTP/1.1 206 Partial Content
+-- > < Transfer-Encoding: chunked
+-- > < Date: Tue, 30 Jan 2018 12:45:17 GMT
+-- > < Server: Warp/3.2.13
+-- > < Content-Type: application/json;charset=utf-8
+-- > < Accept-Ranges: name
+-- > < Content-Range: name Yellow..Purple
+-- > < Next-Range: name Purple;limit 10;offset 1;order desc
+--
+-- The 'Range' header is totally optional, but when provided, it indicates to the server what
+-- parts of the collection is requested. As a reponse and in addition to the data, the server may
+-- provide 3 headers to the client:
+--
+-- - @Accept-Ranges@: A comma-separated list of field upon which a range can be defined
+-- - @Content-Range@: Actual range corresponding to the content being returned
+-- - @Next-Range@: Indicate what should be the next `Range` header in order to retrieve the next range
+--
+-- This allows the client to work in a very _dumb_ mode where it simply consumes data from the server
+-- using the value of the 'Next-Range' header to fetch each new batch of data. The 'Accept-Ranges'
+-- comes in handy to self-document the API telling the client about the available filtering and sorting options
+-- of a resource.
+--
+-- Here's a minimal example used to obtained the previous behavior; Most of the magic happens in the
+-- 'returnRange' function which lift a collection of resources into a Servant handler, computing the
+-- corresponding ranges from the range used to retrieve the resources.
+--
+-- @
+-- -- Resource Type
+--
+-- data Color = Color
+--   { name :: 'String'
+--   , rgb  :: ['Int']
+--   , hex  :: 'String'
+--   } deriving ('Eq', 'Show', 'GHC.Generics.Generic')
+--
+-- colors :: [Color]
+-- colors = [ {- ... -} ]
+--
+-- -- Ranges definitions
+--
+-- instance 'HasPagination' Color "name" where
+--   type 'RangeType' Color "name" = 'String'
+--   'getFieldValue' _ = name
+--
+--
+-- -- API
+--
+-- type API =
+--   "colors"
+--     :> 'Servant.Header' \"Range\" ('Ranges' '["name"] Color)
+--     :> 'Servant.GetPartialContent' '['Servant.JSON'] ('Servant.Headers' ('PageHeaders' '["name"] Color) [Color])
+--
+--
+-- -- Application
+--
+-- defaultRange :: 'Range' "name" 'String'
+-- defaultRange =
+--   'getDefaultRange' ('Data.Proxy.Proxy' @Color) 'Data.Maybe.Nothing'
+--
+-- server :: 'Servant.Server.Server' API
+-- server mrange = do
+--   let range =
+--         'Data.Maybe.fromMaybe' defaultRange (mrange >>= 'extractRange')
+--
+--   'returnRange' range ('applyRange' range colors)
+--
+-- main :: 'IO' ()
+-- main =
+--   'Network.Wai.Handler.Warp.run' 1337 ('Servant.Server.serve' ('Data.Proxy.Proxy' @API) server)
+-- @
+
+{-# LANGUAGE ConstraintKinds     #-}
+{-# LANGUAGE GADTs               #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeApplications    #-}
+{-# LANGUAGE TypeFamilies        #-}
 
 module Servant.Pagination
   (
   -- * Types
-    Range(..)
+  Ranges
+  , Range(..)
   , RangeOrder(..)
   , AcceptRanges (..)
   , ContentRange (..)
-  , NextRange (..)
   , PageHeaders
-  , TotalCount
+  , IsRangeType
 
   -- * Declare Ranges
-  , FromRange(..)
-  , FromHttpApiData(..)
-  , FromRangeOptions(..)
+  , HasPagination(..)
+  , RangeOptions(..)
   , defaultOptions
-  , defaultRange
 
   -- * Use Ranges
-  , HasPagination(..)
+  , extractRange
+  , returnRange
   , applyRange
-
-  -- * Combine Ranges
-  , (:|:)(..)
   ) where
 
-import           Data.List                   (filter, find)
-import           Data.Maybe                  (fromMaybe, listToMaybe)
-import           Data.Proxy                  (Proxy (..))
-import           Data.Semigroup              ((<>))
-import           Data.Text                   (Text)
-import           GHC.Generics                (Generic)
-import           GHC.TypeLits                (KnownSymbol, Symbol, symbolVal)
-import           Numeric.Natural             (Natural)
+import           Data.List      (filter, find)
+import           Data.Maybe     (listToMaybe)
+import           Data.Proxy     (Proxy (..))
+import           Data.Semigroup ((<>))
+import           Data.Text      (Text)
+import           GHC.Generics   (Generic)
+import           GHC.TypeLits   (KnownSymbol, Symbol, symbolVal)
 import           Servant
 
-import qualified Data.List                   as List
-import qualified Data.Text                   as Text
+import qualified Data.List      as List
+import qualified Data.Text      as Text
 import qualified Safe
-
-import           Servant.Pagination.Internal
 
 
 --
 -- TYPES
 --
 
--- An actual Range parsed from a `Range` header. A Range
-data Range (field :: Symbol) typ = Range
-  { rangeValue  :: Maybe typ   -- ^ The value of that field, beginning of the range
+-- | Set of constraints that must apply to every type target of a Range
+type IsRangeType a =
+  ( Show a
+  , Ord a
+  , Eq a
+  , FromHttpApiData a
+  , ToHttpApiData a
+  )
+
+-- | A type to specify accepted Ranges via the @Range@ HTTP Header. For example:
+--
+-- @
+-- type API = "resources"
+--   :> 'Servant.Header' \"Range\" ('Ranges' '["created_at"] Resource)
+--   :> 'Servant.Get' '['Servant.JSON'] [Resource]
+-- @
+data Ranges :: [Symbol] -> * -> * where
+  Lift   :: Ranges fields resource -> Ranges (y ': fields) resource
+  Ranges
+    :: HasPagination resource field
+    => Range field (RangeType resource field)
+    -> Ranges (field ': fields) resource
+
+-- | An actual 'Range' instance obtained from parsing / to generate a @Range@ HTTP Header.
+data Range (field :: Symbol) (a :: *) =
+  (KnownSymbol field, IsRangeType a) => Range
+  { rangeValue  :: Maybe a     -- ^ The value of that field, beginning of the range
   , rangeLimit  :: Int         -- ^ Maximum number of resources to return
   , rangeOffset :: Int         -- ^ Offset, number of resources to skip after the starting value
   , rangeOrder  :: RangeOrder  -- ^ The order of sorting (ascending or descending)
-  } deriving (Eq, Show, Generic)
+  , rangeField  :: Proxy field -- ^ Actual field this Range actually refers to
+  }
 
-instance Functor (Range field) where
-  fmap f r =
-    r { rangeValue = f <$> rangeValue r }
 
-instance (ToHttpApiData typ, KnownSymbol field) => ToHttpApiData (Range field typ) where
-  toUrlPiece Range{..} =
-    Text.pack (symbolVal (Proxy :: Proxy field))
+-- | Extract a 'Range' from a 'Ranges'
+class ExtractRange (fields :: [Symbol]) (field :: Symbol) where
+  -- | Extract a 'Range' from a 'Ranges'. Works like a safe 'read', trying to coerce a 'Range' instance to
+  -- an expected type. Type annotation are most likely necessary to remove ambiguity. Note that a 'Range'
+  -- can only be extrated to a type bound by the allowed 'fields' on a given 'resource'.
+  --
+  -- @
+  -- extractDateRange :: 'Ranges' '["created_at", "name"] Resource -> 'Range' "created_at" 'Data.Time.Clock.UTCTime'
+  -- extractDateRange =
+  --   'extractRange'
+  -- @
+  extractRange
+    :: HasPagination resource field
+    => Ranges fields resource                         -- ^ A list of accepted Ranges for the API
+    -> Maybe (Range field (RangeType resource field)) -- ^ A Range instance of the expected type, if it matches
+
+instance ExtractRange (field ': fields) field where
+  extractRange (Ranges r) = Just r
+  extractRange (Lift _)   = Nothing
+
+instance {-# OVERLAPPABLE #-} ExtractRange fields field => ExtractRange (y ': fields) field where
+  extractRange (Ranges _) = Nothing
+  extractRange (Lift r)   = extractRange r
+
+
+-- | Put a 'Range' in a 'Ranges'
+class PutRange (fields :: [Symbol]) (field :: Symbol) where
+  putRange
+    :: HasPagination resource field
+    => Range field (RangeType resource field)
+    -> Ranges fields resource
+
+instance PutRange (field ': fields) field where
+  putRange = Ranges
+
+instance {-# OVERLAPPABLE #-} (PutRange fields field) => PutRange (y ': fields) field where
+  putRange = Lift . putRange
+
+
+instance ToHttpApiData (Ranges fields resource) where
+  toUrlPiece (Lift range) =
+    toUrlPiece range
+
+  toUrlPiece (Ranges Range{..}) =
+    Text.pack (symbolVal rangeField)
     <> maybe "" (\v -> " " <> toUrlPiece v) rangeValue
     <> ";limit "  <> toUrlPiece rangeLimit
     <> ";offset " <> toUrlPiece rangeOffset
     <> ";order "  <> toUrlPiece rangeOrder
 
-instance (KnownSymbol field) => ToAcceptRanges (Range field typ) where
-  toAcceptRanges _ =
-    Text.pack (symbolVal (Proxy :: Proxy field))
+
+instance FromHttpApiData (Ranges '[] resource) where
+  parseUrlPiece _ =
+    Left "Invalid Range"
+
+instance
+  ( FromHttpApiData (Ranges fields resource)
+  , HasPagination resource field
+  , KnownSymbol field
+  , IsRangeType (RangeType resource field)
+  ) => FromHttpApiData (Ranges (field ': fields) resource) where
+  parseUrlPiece txt =
+    let
+      RangeOptions{..} = getRangeOptions (Proxy @resource) (Proxy @field)
+
+      toTuples =
+        filter (/= "") . Text.splitOn (Text.singleton ' ')
+
+      args =
+        map toTuples $ Text.splitOn (Text.singleton ';') txt
+
+      field =
+        Text.pack $ symbolVal (Proxy @field)
+    in
+      case args of
+        (field' : value) : rest | field == field' -> do
+          opts <-
+            traverse parseOpt rest
+
+          range <- Range
+            <$> sequence (fmap parseQueryParam (listToMaybe value))
+            <*> ifOpt "limit"  defaultRangeLimit opts
+            <*> ifOpt "offset" defaultRangeOffset opts
+            <*> ifOpt "order"  defaultRangeOrder opts
+            <*> pure (Proxy @field)
+
+          pure $ Ranges range
+
+        _ ->
+          Lift <$> (parseUrlPiece txt :: Either Text (Ranges fields resource))
+    where
+      parseOpt :: [Text] -> Either Text (Text, Text)
+      parseOpt piece =
+        case piece of
+          [opt, arg] ->
+            pure (opt, arg)
+
+          _ ->
+            Left "Invalid Range Options"
+
+      ifOpt :: FromHttpApiData o => Text -> o -> [(Text, Text)] -> Either Text o
+      ifOpt opt def =
+        maybe (pure def) (parseQueryParam . snd) . find ((== opt) . fst)
 
 
 -- | Define the sorting order of the paginated resources (ascending or descending)
@@ -92,242 +295,167 @@ instance FromHttpApiData RangeOrder where
       _      -> Left "Invalid Range Order"
 
 
--- | Accepted Ranges in the `Accept-Ranges` response's header
-data AcceptRanges range = AcceptRanges
+-- | Type alias to declare response headers related to pagination
+--
+-- @
+-- type MyHeaders =
+--   'PageHeaders' '["created_at"] Resource
+--
+-- type API = "resources"
+--   :> 'Servant.Header' \"Range\" ('Ranges' '["created_at"] Resource)
+--   :> 'Servant.Get' '['Servant.JSON'] ('Servant.Headers' MyHeaders [Resource])
+-- @
+type PageHeaders (fields :: [Symbol]) (resource :: *) =
+  '[ Header "Accept-Ranges" (AcceptRanges fields)
+   , Header "Content-Range" (ContentRange fields resource)
+   , Header "Next-Range"    (Ranges fields resource)
+   ]
 
-instance (ToAcceptRanges a) => ToHttpApiData (AcceptRanges a) where
-  toUrlPiece _ =
-    toAcceptRanges (Proxy :: Proxy a)
+-- | Accepted Ranges in the `Accept-Ranges` response's header
+data AcceptRanges (fields :: [Symbol]) = AcceptRanges
+
+instance (KnownSymbol field) => ToHttpApiData (AcceptRanges '[field]) where
+  toUrlPiece AcceptRanges =
+    Text.pack (symbolVal (Proxy @field))
+
+instance (ToHttpApiData (AcceptRanges (f ': fs)), KnownSymbol field) => ToHttpApiData (AcceptRanges (field ': f ': fs)) where
+  toUrlPiece AcceptRanges =
+    Text.pack (symbolVal (Proxy @field)) <> "," <> toUrlPiece (AcceptRanges @(f ': fs))
 
 
 -- | Actual range returned, in the `Content-Range` response's header
-data ContentRange range = ContentRange
-  { contentRangeStart :: range
-  , contentRangeEnd   :: range
+data ContentRange (fields :: [Symbol]) resource =
+  forall field. (KnownSymbol field, ToHttpApiData (RangeType resource field)) => ContentRange
+  { contentRangeStart :: RangeType resource field
+  , contentRangeEnd   :: RangeType resource field
+  , contentRangeField :: Proxy field
   }
 
-instance (ToHttpApiData typ, KnownSymbol field) => ToHttpApiData (ContentRange (Range field typ)) where
-  toUrlPiece (ContentRange start end) =
-    Text.pack (symbolVal (Proxy :: Proxy field))
-    <> " "  <> (fromMaybe "" (toUrlPiece <$> rangeValue start))
-    <> ".." <> (fromMaybe "" (toUrlPiece <$> rangeValue end))
+instance ToHttpApiData (ContentRange fields res) where
+  toUrlPiece (ContentRange start end field) =
+    Text.pack (symbolVal field) <> " " <> toUrlPiece start <> ".." <> toUrlPiece end
 
-instance (ToHttpApiData (ContentRange a), ToHttpApiData (ContentRange b)) => ToHttpApiData (ContentRange (a :|: b)) where
-  toUrlPiece (ContentRange (InL sa) (InL ea)) =
-    toUrlPiece (ContentRange sa ea)
-
-  toUrlPiece (ContentRange (InR sb) (InR eb)) =
-    toUrlPiece (ContentRange sb eb)
-
-  toUrlPiece _ =
-    error "impossible"
-
-
--- | Range to provide to retrieve the next batch of resource, in the `Next-Range` response's header
-data NextRange range = NextRange range
-
-instance (ToHttpApiData typ, KnownSymbol field) => ToHttpApiData (NextRange (Range field typ)) where
-  toUrlPiece (NextRange r) =
-    toUrlPiece r
-
-instance (ToHttpApiData (NextRange a), ToHttpApiData (NextRange b)) => ToHttpApiData (NextRange (a :|: b)) where
-  toUrlPiece (NextRange (InL a)) =
-    toUrlPiece (NextRange a)
-
-  toUrlPiece (NextRange (InR b)) =
-    toUrlPiece (NextRange b)
-
-
--- | Type alias to declare response headers related to pagination
-type PageHeaders range =
-  '[ Header "Accept-Ranges" (AcceptRanges range)
-   , Header "Content-Range" (ContentRange range)
-   , Header "Next-Range"    (NextRange range)
-   , Header "Total-Count"   Natural
-   ]
-
-
---
--- DECLARE RANGES
---
-
--- | Default values to apply when parsing a Range
-data FromRangeOptions  = FromRangeOptions
-  { defaultRangeLimit  :: Int
-  , defaultRangeOffset :: Int
-  , defaultRangeOrder  :: RangeOrder
-  } deriving (Eq, Show)
-
-
--- | Some default options of default values for a Range (limit 100; offset 0; order desc)
-defaultOptions :: FromRangeOptions
-defaultOptions =
-  FromRangeOptions 100 0 RangeDesc
-
-
--- | Some default range based on the default options
-defaultRange :: Maybe a -> FromRangeOptions -> Range field a
-defaultRange val opts =
-  let
-    (FromRangeOptions lim off ord) =
-      opts
-  in
-    Range val lim off ord
-
-
--- | Parse a Range object from a `Range` request's header. Any `Range field typ` and combinations
--- of any `Range field typ` provide instance of this class. It is a signature similar to
--- `parseUrlPiece` from the `FromHttpApiData` class and can be used as a drop-in replacement to
--- define instance of this class.
---
--- > type MyRange = Range "created_at" UTCTime
--- >
--- > instance FromHttpApiData UTCTime => FromHttpApiData MyRange where
--- >     parseUrlPiece =
--- >         fromRange defaultOptions
-class FromRange a where
-  parseRange :: FromRangeOptions -> Text -> Either Text a
-
-instance (FromHttpApiData typ, KnownSymbol field) => FromRange (Range field typ) where
-  parseRange FromRangeOptions{..} txt =
-    let
-      toTuples =
-        filter (/= "") . Text.splitOn (Text.singleton ' ')
-
-      args =
-        map toTuples $ Text.splitOn (Text.singleton ';') txt
-
-      field =
-        Text.pack $ symbolVal (Proxy :: Proxy field)
-    in
-      case args of
-        (field' : value) : rest | field == field' -> do
-          opts <-
-            traverse parseOpt rest
-
-          Range
-            <$> sequence (fmap parseQueryParam (listToMaybe value))
-            <*> ifOpt "limit" defaultRangeLimit opts
-            <*> ifOpt "offset"  defaultRangeOffset opts
-            <*> ifOpt "order" defaultRangeOrder opts
-
-        _ ->
-          Left "Invalid Range"
-    where
-      parseOpt :: [Text] -> Either Text (Text, Text)
-      parseOpt piece =
-        case piece of
-          [opt, arg] ->
-            pure (opt, arg)
-
-          _ ->
-            Left "Invalid Range Options"
-
-      ifOpt :: FromHttpApiData o => Text -> o -> [(Text, Text)] -> Either Text o
-      ifOpt opt def =
-        maybe (pure def) (parseQueryParam . snd) . find ((== opt) . fst)
-
-instance {-# Overlappable #-} (FromHttpApiData typ, KnownSymbol field) => FromHttpApiData (Range field typ) where
-  parseUrlPiece =
-    parseRange defaultOptions
-
-
-type TotalCount =
-  Maybe Natural
 
 --
 -- USE RANGES
 --
 
--- | In addition to the `FromHttpApiData` instance, one can provide an instance for this
--- type-class to easily lift a list of response to a Servant handler.
--- By providing a getter to retrieve the value of an actual range from a resource, the
--- `HasPagination` class provides `returnPage` to handle the plumbering of declaring
--- response headers related to pagination.
+-- | Available 'Range' on a given @resource@ must implements the 'HasPagination' type-class.
+-- This class defines how the library can interact with a given @resource@ to access the value
+-- to which a @field@ refers.
 class KnownSymbol field => HasPagination resource field where
   type RangeType resource field :: *
 
-  getRangeField :: Proxy field -> resource -> RangeType resource field
+  -- | Get the corressponding value of a Resource
+  getFieldValue :: Proxy field -> resource -> RangeType resource field
 
-  returnPage_ :: forall m ranges.
-    ( Monad m
-    , (Range field (RangeType resource field)) :<: ranges
-    , ToAcceptRanges ranges
-    , ToHttpApiData (ContentRange ranges)
-    , ToHttpApiData (NextRange ranges)
-    , Ord (RangeType resource field)
-    ) => (Range field (RangeType resource field)) -> [resource] -> m (Headers (PageHeaders ranges) [resource])
-  returnPage_ =
-    returnPage Nothing
-  {-# INLINE returnPage_ #-}
+  -- | Get parsing options for the 'Range' defined on this 'field'
+  getRangeOptions :: Proxy resource -> Proxy field -> RangeOptions
+  getRangeOptions _ _ = defaultOptions
 
-  returnPage :: forall m ranges.
-    ( Monad m
-    , (Range field (RangeType resource field)) :<: ranges
-    , ToAcceptRanges ranges
-    , ToHttpApiData (ContentRange ranges)
-    , ToHttpApiData (NextRange ranges)
-    , Ord (RangeType resource field)
-    ) => TotalCount -> (Range field (RangeType resource field)) -> [resource] -> m (Headers (PageHeaders ranges) [resource])
-  returnPage count range rs = do
-    let field =
-          Proxy :: Proxy field
+  -- | Create a default 'Range' from a value and default 'RangeOptions'. Typical use-case
+  -- is for when no or an invalid 'Range' header was provided.
+  getDefaultRange
+    :: IsRangeType (RangeType resource field)
+    => Proxy resource
+    -> Maybe (RangeType resource field)
+    -> Range field (RangeType resource field)
+  getDefaultRange _ val =
+    let
+      RangeOptions{..} = getRangeOptions (Proxy @resource) (Proxy @field)
+    in Range
+      { rangeValue  = val
+      , rangeLimit  = defaultRangeLimit
+      , rangeOffset = defaultRangeOffset
+      , rangeOrder  = defaultRangeOrder
+      , rangeField  = Proxy @field
+      }
 
-    let boundaries = (,)
-          <$> fmap (getRangeField field) (Safe.headMay rs)
-          <*> fmap (getRangeField field) (Safe.lastMay rs)
+-- | Lift an API response in a 'Monad', typically a 'Servant.Server.Handler'. 'Ranges' headers can be quite cumbersome to
+-- declare and can be deduced from the resources returned and the previous 'Range'. This is exactly what this function
+-- does.
+--
+-- @
+-- myHandler
+--  :: 'Maybe' ('Ranges' '["created_at"] Resource)
+--  -> 'Servant.Server.Handler' ('Servant.Headers' ('PageHeaders' '["created_at"] Resource) [Resource])
+-- myHandler mrange =
+--  let range =
+--        'Data.Maybe.fromMaybe' ('getDefaultRange' ('Data.Proxy.Proxy' @Resource)) (mrange >>= 'extractRange')
+--
+--  'returnRange' range ('applyRange' range resources)
+-- @
+returnRange
+  :: ( Monad m
+     , ToHttpApiData (AcceptRanges fields)
+     , KnownSymbol field
+     , HasPagination resource field
+     , IsRangeType (RangeType resource field)
+     , PutRange fields field
+     )
+  => Range field (RangeType resource field)               -- ^ Actual 'Range' used to retrieve the results
+  -> [resource]                                           -- ^ Resources to returned, fetched from a db or a local store
+  -> m (Headers (PageHeaders fields resource) [resource]) -- ^ Resources embedded in a given 'Monad' (typically a 'Servant.Server.Handler', with pagination headers)
+returnRange Range{..} rs = do
+  let boundaries = (,)
+        <$> fmap (getFieldValue rangeField) (Safe.headMay rs)
+        <*> fmap (getFieldValue rangeField) (Safe.lastMay rs)
 
-    let acceptRanges =
-          addHeader (AcceptRanges :: AcceptRanges ranges)
+  case boundaries of
+    Nothing ->
+      return $ addHeader AcceptRanges $ noHeader $ noHeader rs
 
-    let totalCount =
-          maybe noHeader addHeader count
+    Just (start, end) -> do
+      let nextOffset | rangeValue == Just end = rangeOffset + length rs
+                     | otherwise              = length $ takeWhile ((==) end . getFieldValue rangeField) (reverse rs)
 
-    case boundaries of
-      Nothing ->
-        return $
-          acceptRanges $ noHeader $ noHeader $ totalCount rs
+      let nextRange = putRange Range
+            { rangeValue  = Just end
+            , rangeLimit  = rangeLimit
+            , rangeOffset = nextOffset
+            , rangeOrder  = rangeOrder
+            , rangeField  = rangeField
+            }
 
-      Just (start, end) -> do
-        let rangeStart =
-              liftRange $ (range { rangeValue = Just start } :: Range field (RangeType resource field))
+      let contentRange = ContentRange
+            { contentRangeStart = start
+            , contentRangeEnd   = end
+            , contentRangeField = rangeField
+            }
 
-        let rangeEnd =
-              liftRange $ (range { rangeValue = Just end } :: Range field (RangeType resource field))
+      return $ addHeader AcceptRanges $ addHeader contentRange $ addHeader nextRange rs
 
-        let nextOffset | rangeValue range `compare` Just end == EQ = rangeOffset range + length rs
-                       | otherwise = length $ takeWhile (\r -> getRangeField field r `compare` end == EQ) $ reverse rs
-        
-        let rangeNext =
-              liftRange $ (range { rangeValue = Just end, rangeOffset = nextOffset } :: Range field (RangeType resource field))
-
-        let contentRange =
-              addHeader $ ContentRange
-                { contentRangeStart = rangeStart
-                , contentRangeEnd   = rangeEnd
-                }
-
-        let nextRange =
-              addHeader $ NextRange $ rangeNext
-
-        return
-          $ acceptRanges $ contentRange $ nextRange $ totalCount rs
+-- | Default values to apply when parsing a 'Range'
+data RangeOptions  = RangeOptions
+  { defaultRangeLimit  :: Int         -- ^ Default limit if not provided, default to @100@
+  , defaultRangeOffset :: Int         -- ^ Default offset if not provided, default to @0@
+  , defaultRangeOrder  :: RangeOrder  -- ^ Default order if not provided, default to 'RangeDesc'
+  } deriving (Eq, Show)
 
 
--- | Apply a range to a list of element
-applyRange :: forall b field. (HasPagination b field, Ord (RangeType b field)) => Range field (RangeType b field) -> [b] -> [b]
+-- | Some default options of default values for a Range (limit 100; offset 0; order desc)
+defaultOptions :: RangeOptions
+defaultOptions =
+  RangeOptions 100 0 RangeDesc
+
+
+-- | Helper to apply a 'Range' to a list of values. Most likely useless in practice
+-- as results may come more realistically from a database, but useful for debugging or
+-- testing.
+applyRange
+  :: HasPagination resource field
+  => Range field (RangeType resource field) -- ^ A 'Range' instance on a given @resource@
+  -> [resource]                             -- ^ A full-list of @resource@
+  -> [resource]                             -- ^ The sublist obtained by applying the 'Range'
 applyRange Range{..} =
   let
-    field =
-      Proxy :: Proxy (field :: Symbol)
-
     sortRel =
       case rangeOrder of
         RangeDesc ->
-          \a b -> compare (getRangeField field b) (getRangeField field a)
+          \a b -> compare (getFieldValue rangeField b) (getFieldValue rangeField a)
 
         RangeAsc ->
-          \a b -> compare (getRangeField field a) (getRangeField field b)
+          \a b -> compare (getFieldValue rangeField a) (getFieldValue rangeField b)
 
     dropRel =
       case (rangeValue, rangeOrder) of
@@ -335,10 +463,10 @@ applyRange Range{..} =
           const False
 
         (Just a, RangeDesc) ->
-          (> a) . (getRangeField field)
+          (> a) . getFieldValue rangeField
 
         (Just a, RangeAsc) ->
-          (< a) . (getRangeField field)
+          (< a) . getFieldValue rangeField
   in
       List.take rangeLimit
     . List.drop rangeOffset