+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## v2.0.0 (2018-04-06)
+
+- Review internal implementation and public API (ditch Range combinator to favor type-level
+  list and more discrete footprint). 
+
+- Remove 'Total-Count' header, can still be added on top of the range headers but isn't a Range
+  header so to speak. 
+
+- Extend haddock documentation to be more user-friendly
+
+
 ## v1.0.0 (2018-02-06)
 
 - Initial release